test: 테스트 코드 작성

+++ src/main/java/dto/PlayerDto.java
@@ -28,7 +28,7 @@ public class PlayerDto {
         return sumOfCards;
     }
 
-    public double getMoney() {
+    public int getMoney() {
         return money;
     }
 }
